reviews from pierrick, add two components for ui
