+++ app/src/features/FhirResourceTree/TreeItem.tsx
@@ -1,26 +1,20 @@
 import React, { useState } from "react";
 
-import { Icon, MaybeElement } from "@blueprintjs/core";
 import { IconName, IconNames } from "@blueprintjs/icons";
-import {
-  IconButton as MuiIconButton,
-  makeStyles,
-  Theme,
-  Typography,
-} from "@material-ui/core";
+import { makeStyles } from "@material-ui/core";
 import { TreeItem as MuiTreeItem } from "@material-ui/lab";
-import clsx from "clsx";
 
 import useFhirResourceTreeData from "common/hooks/useFhirResourceTreeData";
 
 import { ElementNode } from "./resourceTreeSlice";
+import TreeItemLabel from "./TreeItemLabel";
 
 type TreeItemProps = {
   elementNode: ElementNode;
   isArrayItem?: boolean;
 };
 
-const useStyle = makeStyles((theme: Theme) => ({
+const useStyle = makeStyles(() => ({
   root: {
     "& .MuiTreeItem-content .MuiTreeItem-label": {
       borderRadius: 4,
@@ -28,67 +22,7 @@ const useStyle = makeStyles((theme: Theme) => ({
       paddingLeft: 10,
     },
   },
-  treeItemContainer: {
-    display: "flex",
-    alignItems: "center",
-    height: "100%",
-    "&:hover": {
-      "& button": {
-        display: "flex",
-      },
-    },
-  },
-  icon: {
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    fill: theme.palette.text.primary,
-  },
-  treeItemTitle: {
-    fontWeight: 500,
-  },
-  treeItemType: {
-    fontWeight: 400,
-    flexGrow: 1,
-    overflow: "hidden",
-    whiteSpace: "nowrap",
-    width: 1,
-    textOverflow: "ellipsis",
-    display: "inline-block",
-  },
-  margin: {
-    marginLeft: theme.spacing(1),
-  },
-  iconRight: {
-    fill: theme.palette.text.secondary,
-  },
-  labelEndIcons: {
-    display: "none",
-    marginRight: 4,
-  },
 }));
-
-type IconButtonProps = {
-  icon: IconName | MaybeElement;
-  onClick: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
-};
-
-const IconButton = ({ icon, onClick }: IconButtonProps): JSX.Element => {
-  const classes = useStyle();
-  return (
-    <MuiIconButton
-      size="small"
-      className={classes.labelEndIcons}
-      onClick={onClick}
-    >
-      <Icon
-        className={clsx(classes.icon, classes.iconRight)}
-        icon={icon}
-        iconSize={15}
-      />
-    </MuiIconButton>
-  );
-};
 
 const TreeItem = ({ elementNode, isArrayItem }: TreeItemProps): JSX.Element => {
   const classes = useStyle();
@@ -100,12 +34,6 @@ const TreeItem = ({ elementNode, isArrayItem }: TreeItemProps): JSX.Element => {
     { id: elementNode.type ?? "", nodeId: elementNode.id },
     { skip: !isComplex || !hasExpanded || elementNode.isArray }
   );
-
-  const handleIconButtonClick = (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
-    e.stopPropagation();
-  };
 
   const handleIconClick = () => {
     setHasExpanded(true);
@@ -130,53 +58,22 @@ const TreeItem = ({ elementNode, isArrayItem }: TreeItemProps): JSX.Element => {
     }
   } else {
     iconName =
-      elementNode.name === "extension"
+      elementNode.type === "Extension"
         ? IconNames.CODE_BLOCK
         : (iconName = IconNames.LAYERS);
   }
-
-  const LabelContent = (): JSX.Element => {
-    return (
-      <div className={classes.treeItemContainer}>
-        {iconName && (
-          <Icon className={classes.icon} icon={iconName} iconSize={15} />
-        )}
-        <Typography
-          className={clsx(classes.margin, classes.treeItemTitle)}
-          display="inline"
-          color="textPrimary"
-        >
-          {elementNode.name}
-        </Typography>
-        <Typography
-          display="inline"
-          variant="subtitle2"
-          color="textSecondary"
-          className={clsx(classes.margin, classes.treeItemType)}
-        >
-          {elementNode.type}
-        </Typography>
-        {elementNode.isArray && (
-          <IconButton icon={IconNames.ADD} onClick={handleIconButtonClick} />
-        )}
-        {isArrayItem && (
-          <IconButton icon={IconNames.TRASH} onClick={handleIconButtonClick} />
-        )}
-        {elementNode.kind === "complex" && !elementNode.isArray && (
-          <IconButton
-            icon={IconNames.CODE_BLOCK}
-            onClick={handleIconButtonClick}
-          />
-        )}
-      </div>
-    );
-  };
 
   return (
     <MuiTreeItem
       nodeId={elementNode.id}
       classes={{ root: classes.root }}
-      label={<LabelContent />}
+      label={
+        <TreeItemLabel
+          iconName={iconName}
+          isArrayItem={isArrayItem}
+          elementNode={elementNode}
+        />
+      }
       onIconClick={handleIconClick}
       onLabelClick={handleLabelClick}
     >
